backend/requirements.txt
gunicorn
psycopg2
Django==2.1.1
django-cors-headers==2.4.0
django-rest-api==0.1.5
django-rest-framework==0.1.0
djangorestframework==3.8.2
mido
scipy
numpy